Move costType component up a level
