+++ src/routes/components/costType/costType.scss
