+++ src/routes/settings/calculations/calculations.tsx
@@ -4,8 +4,8 @@ import React from 'react';
 import type { WrappedComponentProps } from 'react-intl';
 import { injectIntl } from 'react-intl';
 import { connect } from 'react-redux';
+import { CostType } from 'routes/components/costType';
 import { Currency } from 'routes/components/currency';
-import { CostType } from 'routes/views/components/costType';
 import type { FetchStatus } from 'store/common';
 import { createMapStateToProps } from 'store/common';
 import { settingsActions, settingsSelectors } from 'store/settings';
