+++ src/routes/views/details/awsDetails/detailsHeader.tsx
@@ -13,8 +13,8 @@ import React from 'react';
 import type { WrappedComponentProps } from 'react-intl';
 import { injectIntl } from 'react-intl';
 import { connect } from 'react-redux';
+import { CostType } from 'routes/components/costType';
 import { Currency } from 'routes/components/currency';
-import { CostType } from 'routes/views/components/costType';
 import { GroupBy } from 'routes/views/components/groupBy';
 import { filterProviders } from 'routes/views/utils/providers';
 import type { FetchStatus } from 'store/common';
