+++ src/routes/views/details/components/breakdown/breakdownHeader.tsx
@@ -13,10 +13,10 @@ import { injectIntl } from 'react-intl';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { routes } from 'routes';
+import { CostType } from 'routes/components/costType';
 import { Currency } from 'routes/components/currency';
 import { ComputedReportItemValueType } from 'routes/views/components/charts/common';
 import { CostDistribution } from 'routes/views/components/costDistribution';
-import { CostType } from 'routes/views/components/costType';
 import { TagLink } from 'routes/views/details/components/tag';
 import { getGroupByCostCategory, getGroupByOrgValue, getGroupByTagKey } from 'routes/views/utils/groupBy';
 import { createMapStateToProps } from 'store/common';
