+++ src/routes/views/explorer/explorerHeader.tsx
@@ -15,9 +15,9 @@ import React from 'react';
 import type { WrappedComponentProps } from 'react-intl';
 import { injectIntl } from 'react-intl';
 import { connect } from 'react-redux';
+import { CostType } from 'routes/components/costType';
 import { Currency } from 'routes/components/currency';
 import { CostDistribution } from 'routes/views/components/costDistribution';
-import { CostType } from 'routes/views/components/costType';
 import { GroupBy } from 'routes/views/components/groupBy';
 import { Perspective } from 'routes/views/components/perspective';
 import type { Filter } from 'routes/views/utils/filter';
