+++ src/routes/views/overview/overview.tsx
@@ -26,11 +26,11 @@ import React from 'react';
 import type { WrappedComponentProps } from 'react-intl';
 import { injectIntl } from 'react-intl';
 import { connect } from 'react-redux';
+import { CostType } from 'routes/components/costType';
 import { Currency } from 'routes/components/currency';
 import { Loading } from 'routes/state/loading';
 import NoData from 'routes/state/noData/noData';
 import { NoProviders } from 'routes/state/noProviders';
-import { CostType } from 'routes/views/components/costType';
 import { Perspective } from 'routes/views/components/perspective';
 import { AwsDashboard } from 'routes/views/overview/awsDashboard';
 import { AwsOcpDashboard } from 'routes/views/overview/awsOcpDashboard';
